Move all toolbar code to separate classes
