+++ Editor/CustomToolbar/Scripts/CustomToolbarRight.cs
@@ -7,27 +7,12 @@ namespace UnityToolbarExtender {
 
 	[InitializeOnLoad]
 	public static class CustomToolbarRight {
-		
 		private static CustomToolbarSetting setting;
-		
-		private static GUIContent recompileBtn;
-		private static GUIContent reserializeSelectedBtn;
-		private static GUIContent reserializeAllBtn;
-		private static int selectedFramerate = 60;
 
 		static CustomToolbarRight() {
 			setting = CustomToolbarSetting.GetOrCreateSetting();
 			
 			ToolbarExtender.RightToolbarGUI.Add(OnToolbarGUI);
-
-			recompileBtn = EditorGUIUtility.IconContent("WaitSpin05");
-			recompileBtn.tooltip = "Recompile";
-
-			reserializeSelectedBtn = EditorGUIUtility.IconContent("Refresh");
-			reserializeSelectedBtn.tooltip = "Reserialize Selected Assets";
-
-			reserializeAllBtn = EditorGUIUtility.IconContent("P4_Updating");
-			reserializeAllBtn.tooltip = "Reserialize All Assets";
 		}
 
 		static void OnToolbarGUI() {
@@ -38,30 +23,7 @@ namespace UnityToolbarExtender {
 			for (++i; i < setting.elements.Count; ++i)
 				setting.elements[i].DrawInToolbar();
 
-			DrawRecompileButton();
-			DrawReserializeSelected();
-			DrawReserializeAll();
-
 			GUILayout.FlexibleSpace();
-		}
-
-		static void DrawRecompileButton() {
-			if (GUILayout.Button(recompileBtn, ToolbarStyles.commandButtonStyle)) {
-				UnityEditor.Compilation.CompilationPipeline.RequestScriptCompilation();
-				Debug.Log("Recompile");
-			}
-		}
-
-		static void DrawReserializeSelected() {
-			if (GUILayout.Button(reserializeSelectedBtn, ToolbarStyles.commandButtonStyle)) {
-				ForceReserializeAssetsUtils.ForceReserializeSelectedAssets();
-			}
-		}
-
-		static void DrawReserializeAll() {
-			if (GUILayout.Button(reserializeAllBtn, ToolbarStyles.commandButtonStyle)) {
-				ForceReserializeAssetsUtils.ForceReserializeAllAssets();
-			}
 		}
 	}
 }
